Refactor model classes and extend profile
