+++ ProjectPortfolio/Model/Entity/Profile.cs
@@ -4,6 +4,13 @@ namespace Model.Entity;
 
 public class Profile : AEntity
 {
+    public string FirstName { get; set; }
+    public string LastName { get; set; }
+    public string Title { get; set; }
+    public string Location { get; set; }
+    public string ContactEmail { get; set; }
+    public string Phone { get; set; }
+    public string SelfDescription { get; set; }
     public List<EmploymentRecord> EmploymentHistory { get; set; } = new();
     public List<EducationRecord> EducationHistory { get; set; } = new();
     public List<Certification> Certifications { get; set; } = new();
