+++ README.md
@@ -1,1 +1,16 @@
 # ProjectPortfolio2025
+
+This repository contains a sample .NET project portfolio. The `Model` project defines entities that describe a person's professional profile, employment history, education and skills.
+
+## Profile model
+
+The `Profile` class includes identity details useful for rendering a portfolio site:
+
+- `FirstName` and `LastName`
+- `Title` (professional headline)
+- `Location`
+- `ContactEmail` and `Phone`
+- `SelfDescription` for a short biography
+- Lists of employment records, education records, certifications and personal projects
+
+Each project has been updated so that every class resides in its own file.
